Add files via upload
Bag of Words vectorization applied to (artificial) procedure list with names. Frequency information is extracted and graphs are built for the vectorized data. Lastly clustering is performed to categorize people, as evidence of radial regions seen in graphs. The dimensionality is high and therefore it is reduced using PCA. Lastly, mean shifting is used to estimate the cluster number and classify people.